requirements.txt
django
django-cors-headers
djangorestframework
djoser
dateutils
psycopg2-binary
django-multiselectfield